refactor test.
