+++ .gitignore
@@ -9,6 +9,7 @@
 .history
 .svn/
 migrate_working_dir/
+coverage/
 
 # IntelliJ related
 *.iml
